LWC: Replace Number with Double class

+++ app/src/jvmMain/kotlin/com/saurabhsandav/core/ui/stats/studies/PNLByMonthChartStudy.kt
@@ -65,7 +65,7 @@ internal class PNLByMonthChartStudy(
                                 month = localDate.monthNumber,
                                 day = localDate.dayOfMonth,
                             ),
-                            value = bigDecimal,
+                            value = bigDecimal.toDouble(),
                         )
                     }
             }
